deleting the BlogRoll test file
